[Refactor] Move ConsoleEx to its own file
