+++ Source/Libraries/NetCore/NetCore.csproj
@@ -67,6 +67,7 @@
     <Compile Include="SafeJsonTypeSerialization\UntypedToTypedValueConverter.cs" />
     <Compile Include="AllSpec.cs" />
     <Compile Include="Connector.cs" />
+    <Compile Include="ConsoleEx.cs" />
     <Compile Include="NetCore_Extensions.cs" />
     <Compile Include="DebugInfo\CloudDebug_Form.cs">
       <SubType>Form</SubType>
